refactor graph test
