+++ conans/test/integration/graph/core/alias_test.py
@@ -1,7 +1,7 @@
 import six
 
 from conans.errors import ConanException
-from conans.test.integration.graph.graph_manager_base import GraphManagerTest
+from conans.test.integration.graph.core.graph_manager_base import GraphManagerTest
 
 
 class AliasTest(GraphManagerTest):
