+++ conans/test/integration/graph/core/cross_build/_base_test_case.py
@@ -10,7 +10,7 @@ from conans.client.recorder.action_recorder import ActionRecorder
 from conans.model.graph_info import GraphInfo
 from conans.model.options import OptionsValues
 from conans.model.ref import ConanFileReference
-from conans.test.integration.graph.graph_manager_base import GraphManagerTest
+from conans.test.integration.graph.core.graph_manager_base import GraphManagerTest
 from conans.test.utils.test_files import temp_folder
 from conans.util.files import save
