+++ conans/test/integration/graph/core/cross_build/breq_of_breq_test.py
@@ -2,7 +2,7 @@ import textwrap
 
 from conans.client.graph.graph import CONTEXT_BUILD, CONTEXT_HOST
 from conans.model.profile import Profile
-from conans.test.integration.cross_building.graph._base_test_case import CrossBuildingBaseTestCase
+from conans.test.integration.graph.core.cross_build._base_test_case import CrossBuildingBaseTestCase
 
 
 class BuildRequireOfBuildRequire(CrossBuildingBaseTestCase):
