+++ conans/test/integration/graph/core/cross_build/build_requires_in_profile_test.py
@@ -5,7 +5,7 @@ from parameterized import parameterized
 from conans.client.graph.graph import CONTEXT_BUILD, CONTEXT_HOST
 from conans.model.profile import Profile
 from conans.model.ref import ConanFileReference
-from conans.test.integration.cross_building.graph._base_test_case import CrossBuildingBaseTestCase
+from conans.test.integration.graph.core.cross_build._base_test_case import CrossBuildingBaseTestCase
 
 
 class BuildRequiresInProfileExample(CrossBuildingBaseTestCase):
