+++ conans/test/integration/graph/core/cross_build/protobuf_test.py
@@ -3,7 +3,7 @@ import textwrap
 from conans.client.graph.graph import CONTEXT_BUILD, CONTEXT_HOST
 from conans.model.profile import Profile
 from conans.model.ref import ConanFileReference
-from conans.test.integration.cross_building.graph._base_test_case import CrossBuildingBaseTestCase
+from conans.test.integration.graph.core.cross_build._base_test_case import CrossBuildingBaseTestCase
 
 
 class ProtobufTest(CrossBuildingBaseTestCase):
