+++ conans/test/integration/graph/core/cross_build/protoc_basic_test.py
@@ -4,7 +4,7 @@ from parameterized import parameterized
 
 from conans.client.graph.graph import CONTEXT_BUILD, CONTEXT_HOST
 from conans.model.profile import Profile
-from conans.test.integration.cross_building.graph._base_test_case import CrossBuildingBaseTestCase
+from conans.test.integration.graph.core.cross_build._base_test_case import CrossBuildingBaseTestCase
 
 
 class ClassicProtocExampleBase(CrossBuildingBaseTestCase):
