+++ conans/test/integration/graph/core/cross_build/protoc_with_gtest_test.py
@@ -2,7 +2,7 @@ import textwrap
 
 from conans.client.graph.graph import CONTEXT_HOST, CONTEXT_BUILD
 from conans.model.profile import Profile
-from conans.test.integration.cross_building.graph.protoc_basic_test import ClassicProtocExampleBase
+from conans.test.integration.graph.core.cross_build.protoc_basic_test import ClassicProtocExampleBase
 
 
 class ProtocWithGTestExample(ClassicProtocExampleBase):
@@ -19,7 +19,7 @@ class ProtocWithGTestExample(ClassicProtocExampleBase):
             version = "testing"
 
             settings = "os"
-            
+
             def requirements(self):
                 self.requires("protobuf/testing@user/channel")
